requirements.txt
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
matplotlib==3.8.2
dashscope==1.14.0  # 通义千问API依赖（若用GPT则替换为openai==1.13.3）
tenacity==8.2.3    # LLM调用重试
python-dotenv==1.0.1  # 可选：安全管理API密钥